feat: add font to project
