+++ package.json
@@ -17,6 +17,7 @@
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
+    "file-loader": "^6.2.0",
     "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.5.1",
     "husky": "^8.0.3",
@@ -25,6 +26,7 @@
     "sass-loader": "^13.2.2",
     "sass-resources-loader": "^2.2.5",
     "style-loader": "^3.3.2",
+    "url-loader": "^4.1.1",
     "webpack": "^5.82.1",
     "webpack-cli": "^5.1.1",
     "webpack-dev-server": "^4.15.0",
