+++ src/index.scss
@@ -1,15 +1,19 @@
+@import url('https://fonts.googleapis.com/css2?family=Genos:ital,wght@0,100..900;1,100..900&family=Inter:wght@100;200;300;400;500;600;700&family=Lemon&family=Russo+One&display=swap');
 @import './styles/vars.scss';
 @import './styles/initial.scss';
 @import './styles/modal.scss';
 
 * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
 }
 
 body {
-    background-color: rgb(228, 180, 162);
-    max-width: 1440px;
-    margin: 0 auto;
+  background-color: rgb(228, 180, 162);
+  max-width: 1440px;
+  margin: 0 auto;
+  font-family: 'Russo One', sans-serif;
+  font-weight: 400;
+  font-style: normal;
 }
