+++ webpack.config.js
@@ -12,6 +12,12 @@ module.exports = {
   module: {
     rules: [
       {
+        test: /\.(woff|woff2|eot|ttf|otf)$/,
+        use: [
+          'file-loader',
+        ],
+      },
+      {
         test: /\.html$/i,
         loader: 'html-loader',
         options: { minimize: false },
